fix(scripts): pass footer scripts context
